taodata/requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
get==2018.11.19
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
lxml==4.2.5
parsel==1.5.0
Pillow==5.2.0
post==2018.11.20
public==2018.11.20
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pymongo==3.5.1
PyMySQL==0.9.2
pyOpenSSL==18.0.0
query-string==2018.11.20
queuelib==1.5.0
redis==2.10.6
requests==2.20.1
schedule==0.5.0
Scrapy==1.5.1
scrapy-redis==0.6.8
selenium==3.6.0
service-identity==17.0.0
six==1.11.0
Twisted==18.7.0
urllib3==1.24.1
w3lib==1.19.0
zope.interface==4.5.0
